Moved electronic band gap definition to subfolder properties/
